pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^5.0"
python = "^3.10"
redis = "5.0.7"
docker = "7.1.0"
flask = "^3.1.0"
twilio = "^9.5.2"
pytesseract = "^0.3.13"
pillow = "^11.1.0"
openai = "^1.72.0"
requests = "^2.32.3"
google-generativeai = "^0.8.4"
pypdf2 = "^3.0.1"
[tool.poetry.dev-dependencies]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"